+++ components/home-carousel.jsx
@@ -3,7 +3,7 @@ import {carouselData} from '../data/carousel_data';
 import styles from '@/styles/carousel.module.scss'
 export default function HomeCarousel(){
     return(
-        <Carousel className={styles.caroWrapper} arrows dots dotPosition='bottom' autoplay={{ dotDuration: true }} autoplaySpeed={4500} infinite>
+        <Carousel className={styles.caroWrapper} arrows dots dotPosition='top' autoplay={{ dotDuration: true }} autoplaySpeed={4500} infinite>
           {carouselData.map((entry, index) =>
               <div key={index}>
                 <img className={styles.caroImage}