+++ changelog.md
@@ -2,3 +2,13 @@
 - initial CNA
 - inital page setup - navbar, footer, text only
 - link to orig repo
+
+# v1.0.0
+- release with portfolio page active
+
+# v1.1.0
+- added carousel to homepage
+- mobile navbar styling
+- fixing other mobile styling
+- font-family and text cleanup
+- flex tags in "about" page